refactor(analytics): extract scope/sample math into AnalyticsScopeMath (P10.1)
Second cut of decomposing the ChampionAnalyticsComputeService god-file. The shared,
stateless rank-tier-scope + sample-size math — RankTierScope + ParseRankTierScope,
ApplyRankTierScopeToParticipants, ResolveEffectiveSampleSize, ComputeWilsonLowerBound,
ResolvePlatformsForRegion, ScopeTokenOf — moves verbatim into a single AnalyticsScopeMath
static class, shared cleanly by the raw + stats read paths (which previously reached it
via private members on the partial). Context-coupled callers (GetAdaptiveMinimumGames-
RequiredAsync, BuildScopedMatchIdQuery) stay and route through it.

Pure, behavior-preserving move — no logic changed. ChampionAnalyticsComputeService.cs
1300 -> 1217 lines; ~28 call sites re-pointed across both partials; 0 leftover decls.

Gate: solution builds clean; full Service.Core (183, incl. all four precompute
equivalence-test suites) + WebAPI (40) green. No schema change.
